+++ packages/core/styles/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Nimbus Styles deprive all styles needed to build nimbus components.
 
+## 2025-12-15 `9.47.0`
+
+#### 🎉 New features
+
+- `Table`: Allow table layout to be controlled by the column layout. ([#393](https://github.com/TiendaNube/nimbus-design-system/pull/393) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-12-19 `9.46.1`
 
 #### 🐛 Bug fixes